WIP new version of ADT
Hopefully this will give us better automation of combinators.  The idea
is to write the combinators in terms of index sets, so that we don't
need to worry about duplication or renaming or anything like that.  Then
the burden of matching up methods is on the automation.  I'm not
totally sure if this will end up working well...