Update SAT with some suggestsions from Adam:
On Thu, Nov 28, 2013 at 9:59 AM, Adam Chlipala <adamc@csail.mit.edu> wrote:
> On 11/27/2013 07:07 PM, Jason Gross wrote:
> > I've taken an hour or so and implemented most of what we talked about
> > yesterday, and pushed to the adt-synthesis repo.  It's at the top of SAT.v.
>
> Neato!  Some comments:
>
> Can we commit to writing code that will work in the latest mainstream Coq
> release (i.e., 8.4pl2 at the moment)?  There are some practical reasons for
> this choice that we can debate, but I hope it's low enough cost for both of
> us to be a no-brainer.  8.4pl2 gracefully ignores the [Asymmetric Patterns]
> and [Universe Polymorphism] options, but of course things will go very wrong
> if you write any code that depends on either setting.  I'd rather remove
> those commands.

Sure.  How much do you care about easily fixed typos?  (When I develop on
cagnode, which happens less and less now that I have a good laptop, it's
slightly easier for me to not keep switching between Coq 8.4pl2 and HoTT/coq.
I think I know the features well enough to not write code that essentially
depends on them, but I might, e.g., pass too many underscores to patterns in
pattern matches ([Set Asymmetric Patterns] should help with this).  The problem
with just removing the commands is that then the code won't work with HoTT/coq
at all; [Set Asymmetric Patterns] is kind-of like the compatibility mode flag
for the [match] construct, and the current version of HoTT/coq doesn't have a
compatibility mode for universe polymorphism (which will be in trunk, when the
code finally gets merged), and so I need [Set Universe Polymorphism] to get
[Inductive]s to be sort-polymorphic at all in HoTT/coq).

> Maybe [get_vars] should return [Ensemble] (or a decidable equivalent with
> [bool] instead of [Prop]) instead of [list].  Do you ever need to iterate
> over it?  Forcing a notion of ordering for an inherently unordered value may
> complicate proofs later on.

Done.

> [subst_vars] looks like it's combining elements of the specification of SAT
> and a simplification operation used by a particular implementation.  IMO, it
> would be nicer to have a simpler function that does exactly what is needed to
> specify SAT.  (Now I see that there is no specification here yet.)

I've added [denote_formula] and [is_satisfiable].

> [sat] looks unnecessarily complicated by not making [var] and its decidable
> equality section [Variable]s.  For instance, it seems punitive to pick [x0]
> out of a [sigT].  I really think the right strategic move in this project is
> to avoid dependent types whenever possible.

Done.  I guess it's rare that you'll have decidable equality for the relevant
subtype, but not for the entire type.

> Something else suspicious about the [sat] definition: Isn't it lacking a base
> case?  The [Pick] looks doomed when the input formula [f] is closed.

The base case is part of [subst_vars]; it returns a boolean value on closed
formulas and a [formula] in other cases.  The type of [formula] has no
substituted variables, so there's no such thing as a closed [formula].

It might also be nicer to add [True] as a [formula] constructor, to avoid the
need for sum types in the function to substitute in formulas, and to avoid the
verbose case analysis in [sat].  Instead, the base case of [sat] (for closed
formulas) could check whether the formula is [True] or [False].  (We'd try to
maintain the invariant that closed formulas are one or the other,
syntactically.)

I've made a separate section ([Module maybe_closed], below [Module open_only])
with this version of formulas.  I'll probably pick one or the other soon.  I'm
open to input about which you think works better.  (I like the simplified
implementation of [sat] for the [maybe_closed] version.)  Why do we want to
maintain the invariant that a closed formula is either [True] or [False]?  It
would simplify [subst_vars] to not have this.

> It might be nice to implement a short-circuiting [or] combinator for [Comp],
> to use in, e.g., the place in [sat] with [orb].  Of course the performance
> benefit would be nice, but it should also lead to more readable (better
> abstracted) code.

Done, but it's not clear to me that it's any better abstraction-wise.  (And it
feels poor to me to make Comp have a special case for bool, from an abstraction
stand-point.)

> Did you consider proving the usual monad laws for [computes_to] as a sanity
> check?

No.  But I will refactor the notation to be of any monad, and prove the laws
for [comp] soon.  (Did you mean [comp] rather than [computes_to]?)

> > how to factor things through adts (presumably the computation traces?)

> Therein lies the research. :)