Finish porting what was finished of ADTs to Comp
Everything up to and including the example of refinement using ADTs now uses
Comp.  I'm not exactly sure what comes next, but I have a bit of an idea about
what to explore.  I think what comes next is cleaning up the example of
refinement, and then working on porting the examples in Computation.v to use
ADTs, and then find some more algorithms to play with.