Make progress on the mutator in min+max
The mutator now calls the two separate mutators from the min&max and
the spec implementation.  However, it has some very unfortunate
constructs, like

  x <- { _ : { x : A | computes_to bar x } | True };
  ...

This is necessary to generate (inside the computation monad) the proof
that the the [x] which [bar] computes to satisfies some properties,
which mention [bar].  I suspect that I'll be able to remove these
constructs later in the refinement process, when I figure out how to
say that you can strip off the proof part of a [sig] representation if
it's only ever used to compute the proof part of a mutator.  But I
don't think I can hide all the details behind the "general delegation
refinement".

I kind-of want to thread the proofs through variants of, e.g., [Pick],
so that we can get [Comp] support for this.  Alternatively, I kind-of
want to build the representation invariants into the ADT structure, so
we don't have to use the proofs in the [Comp].  But there might be
another better way to do this, that I'm not seeing.