Partial work on Min+Max
After much time failing to convince myself to work on this, I've
finally managed to get back to work.  I got mired in Ben's notations,
but, hopefully, he can help me out of my mire.

The idea is the following:

Say we have an ADT A and we want to delegate to an ADT B.  Then we
first refine to repA * repB, where we call both mutators in the
mutators, and ignore B in the observers.  This is the part that I'm
nearly done with.

Then we say that, given any relation R between repA and repB, if all
mutators respect R, then you can refine one observer at a time under
the assumption that that relation holds.  Here is were we'll want
tactic support for rewriting one pick into another pick, and then pick
into call, or something like that.

Finally, we say that if you provide a set of observers implemented
over the [snd] projection, and you provide two sets of mutators, one
of which uses only the [fst] and the other of which uses only the
[snd], then we can erase the [fst] projection.

Post-finally, we turn this into the min+max example.