Full tree-based refinement!
Rewrite the tree-based refinement with fresh PIDs, and verify that the
constraints are indeed respected. Some verifications are still somewhat
cumbersome. I'm also thinking that it would be nice to get a specialized
version of 'simplify with monad laws' that handles just the bind-a-`ret`
case, sinc this is what I most often use it for in refinements.
The proofs in ListBasedRefinement and TreeBasedRefinement could also
probably be synced, there still is some duplication/redundancy between
both.