Prove correction (in a certain sense) of nested trees
An important part of the TreeBasedRefinement example is to swap from a
filter over the list of elements to a find over the tree structure. This
was previously justified by manual proofs relying on the KeyedOnPID
predicate; this commit subsumes these proofs by introducing a
bfind_correct predicate in the Bag interface, and proving this predicate
for both the fmap- and the list-based implementations.

The bag interface now has two extra fields, bfind_matcher and
bfind_correct; bfind_matcher is a boolean function which, used in
conjunction with a search toer to filter over the results of benumerate,
must return the same as calling bfind directly.

bstar is the empty search term (such that find bstar == benumerate (the
bstar_search predicate).