Preliminary work on caching
Introduce a caching-able bag interface. Forseeable difficulties include:
* Our set-based specifications of queries only describe results in terms
of membership, thus allowing duplication in the computed result set. For
this reason, cacheable functions are restricted to morphisms of the
SetEq predicate (ie. the cached function f must verify (forall x, In x
seq1 <-> In x seq2) -> fold f seq1 = fold f seq2. max is an example of
such a function, but other functions such as sum are not. Adding non-
duplication to the specification or to the Bag interface will make
implementation of some indexing systems harder. On the other hand, there
is a case for disallowing spurious repetitions.
* The equivalence relation between aggregated results in the caching
interface is equality, while clients might want to keep traces of things
like 'top n elements', which would require a different equivalence
relation (in this example, Permutation).