Introduce tighter constraints on benumerate by introducing a bcount predicate
Assuming decidable equality on TItem, bags are now required to keep
track of how many times each items appear; the count function is not
required to be efficient though, since it mostly serves to assert
properties about the behaviour of binsert and benumerate.

The two interesting predicates are
  forall item, bcount beq bempty item = 0
and
  bcount beq (binsert container inserted) item =
  bcount beq container item + (if beq item inserted then 1 else 0).
where beq designates decidable equality.

The three existing implementations of bag have been augmented to show
that they conform to this extended interface.