Introduce CountingListBags to speed up NumOrders
NumOrders used to be linear in the number of orders, because it had to
run over the lists of orders to count them. ListBags now keep track of
their size, thus making counting much faster.

Note: This is a very simple implementation, but it lacks generality. A cleaner solution would be to create a Bag wrapper that adds counting on top of an existing bag. The only difficulty is that this requires the comparison of an arbitrary search term to bstar to be decideable; this is trivial in the case of CountingListBags, as search terms are just lists of functions, and bstar is the empty list.