Update parser
Now we match the data structure of [productions] more; instead of
asking for all the splits at once, we only ask for splits one at a
time, and iterate down the list ourselves.  This simplifies the code,
and probably also generalizes very nicely to the dependently typed
case.  It will also make the proofs more feasable.