Finish proving compile_fold_base_sca
The lemma is proved in a very exploratory and manual manner; the new
separation between adts and words makes it necessary too have to lemmas,
one for folds producing ADTs and one for producing words, but I'm hoping
that the two proofs should only differ in a limited number of places,
and that work on automating and cleaning up the first proof will provide
a very similar proof for the second one.

Still TODO: express and prove the other fold lemmas, using fold_base.