Finish fixing most of the parser
It now mostly works with the new StringLike.

I think I should go back to parameterizing it on the type of characters,
though.  This allows the interface to have a grammar in the context that
mentions only the type of characters, and to put the stringlike bits
inside the interface.