Rewrite parser to simpl well-foundedness
The parser no longer needs to assume anything about well-founded
relations on strings or lists of nonterminals; it carries around its own
fuel (which, for correctness, must line up with the length of the string
being parsed, and an upper bound on the list of valid nonterminals).

This means that the parser no longer needs to know anything about the
proofy bits of the string-like type.  With some extra contortions, we
can show that having extra proofy bits doesn't change the behavior of
the parser, and thus that we are free to use the correctness proofs that
come from the string carrying around its proofy bits with it, and the
parser that comes from the string not carrying around any proofy bits.
Not requiring the proofy bits speeds up the (ab)* grammar from around 15
seconds to around 2 seconds, and speeds up the parenthesized numbers
grammar from around 30 minutes to just over a minute!