Use indexed strings earlier, and don't duplicate the length
Pushing the optimization of pairing a string with an offset and a
length back to the initial implementation of the boolean recognizer
allows deduplicating the passing of the length of the string, and
makes it easier to remove spurious allocations in Ocaml code.

There are still some places where we do things with [substring], which
should be removed.

This changeset also allows the use of character ranges in the
definition of a grammar.