More work in progress on parser reflective sound
There's a giant snag, which is that one of the higher-order functions
passed in is unspecified on invalid inputs, which it is sometimes
passed, but only when guarded by an [&&] that makes the return value not
matter.  Except there's a lot of reasoning that goes into proving that
this is the case :/

Also, the partial unfolding of first_index_default is invalid; it's only
valid when the function returns a constant or a nonconstant uniformly;
if sometimes it returns a value and sometimes a term, then we might miss
the first index.