Fixed not_found anomaly in QueryStructure examples.
The (somewhat lame) solution was to replace the body of
fold_heading_hyps with idtac. As far as I can tell, the latest version
of transparent abstract fails to perpetuate some identifiers after
Defined, in this particular case the identifiers attached to cached
headings.  Strings can still be safely cached though, and these seem
to be the major source of performance degradation.